Move [gG]etFirstContainerWithSourceVolume to dev/common package and use it for Podman
