+++ pkg/dev/common/common.go
@@ -1,0 +1,26 @@
+package common
+
+import (
+	"fmt"
+
+	"github.com/devfile/library/pkg/devfile/generator"
+
+	corev1 "k8s.io/api/core/v1"
+)
+
+// GetFirstContainerWithSourceVolume returns the first container that set mountSources: true as well
+// as the path to the source volume inside the container.
+// Because the source volume is shared across all components that need it, we only need to sync once,
+// so we only need to find one container. If no container was found, that means there's no
+// container to sync to, so return an error
+func GetFirstContainerWithSourceVolume(containers []corev1.Container) (string, string, error) {
+	for _, c := range containers {
+		for _, env := range c.Env {
+			if env.Name == generator.EnvProjectsSrc {
+				return c.Name, env.Value, nil
+			}
+		}
+	}
+
+	return "", "", fmt.Errorf("in order to sync files, odo requires at least one component in a devfile to set 'mountSources: true'")
+}
